fixing data type issue of dto(s) in accordance with api structure

+++ app/src/main/java/com/vitaflow/app/data/remote/dto/Servings.kt
@@ -9,7 +9,7 @@ data class Servings(
     @SerializedName("raw")
     val raw: String,
     @SerializedName("size")
-    val size: Int,
+    val size: Double,
     @SerializedName("unit")
     val unit: String
 )
